backend/requirements.txt
flask
flask-cors
torch
transformers
captum
numpy
requests
dnspython
Flask-SQLAlchemy
